+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".presentation.MainActivity"
+            android:name=".main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -27,10 +27,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".presentation.SettingsActivity" />
-        <activity android:name=".presentation.MediaActivity" />
-        <activity android:name=".presentation.SearchActivity" />
-        <activity android:name=".presentation.AudioPlayerActivity" />
+        <activity android:name=".settings.ui.SettingsActivity" />
+        <activity android:name=".media.MediaActivity" />
+        <activity android:name=".search.ui.SearchActivity" />
+        <activity android:name=".player.ui.AudioPlayerActivity" />
 
     </application>
