--- app/src/main/java/com/example/playlistmaker/data/dto/TrackRequest.kt
@@ -1,3 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-data class TrackRequest(val expression: String)
