--- app/src/main/java/com/example/playlistmaker/data/dto/TrackResponse.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-import com.google.gson.annotations.SerializedName
-
-data class TrackResponse(
-    @SerializedName("results") val tracks: List<TrackDto>
-) : Response()
