--- app/src/main/java/com/example/playlistmaker/data/network/NetworkClient.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.data.network
-
-import com.example.playlistmaker.data.dto.Response
-
-interface NetworkClient {
-    suspend fun doRequest(dto: Any): Response
-}
