+++ app/src/main/java/com/example/playlistmaker/media/MediaActivity.kt
@@ -1,5 +1,5 @@
 // MediaActivity.kt
-package com.example.playlistmaker.presentation
+package com.example.playlistmaker.media
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
