+++ app/src/main/java/com/example/playlistmaker/player/ui/AudioPlayerActivity.kt
@@ -1,19 +1,20 @@
-package com.example.playlistmaker.presentation
+package com.example.playlistmaker.player.ui
 import android.media.MediaPlayer
 import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import android.util.Log
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.example.playlistmaker.R
+import com.example.playlistmaker.search.domain.models.Track
+import com.example.playlistmaker.search.ui.KEY_FOR_PLAYLIST
 import com.google.gson.Gson
 import java.text.SimpleDateFormat
-import java.util.*
-import android.os.Handler
-import android.os.Looper
-import android.util.Log
-import com.example.playlistmaker.R
-import com.example.playlistmaker.domain.models.Track
+import java.util.Locale
 
 class AudioPlayerActivity : AppCompatActivity() {
     companion object {
