+++ app/src/main/java/com/example/playlistmaker/search/data/TracksRepository.kt
@@ -1,6 +1,6 @@
-package com.example.playlistmaker.domain.api
+package com.example.playlistmaker.search.data
 
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.search.domain.models.Track
 import com.example.playlistmaker.utils.Resource
 
 interface TracksRepository {
