+++ app/src/main/java/com/example/playlistmaker/search/data/dto/Response.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data.dto
+package com.example.playlistmaker.search.data.dto
 
 open class Response {
     var resultCode: Int = 0
