+++ app/src/main/java/com/example/playlistmaker/search/data/dto/TrackRequest.kt
@@ -1,0 +1,3 @@
+package com.example.playlistmaker.search.data.dto
+
+data class TrackRequest(val expression: String)
