+++ app/src/main/java/com/example/playlistmaker/search/data/network/RetrofitNetworkClient.kt
@@ -1,9 +1,9 @@
-package com.example.playlistmaker.data.network
+package com.example.playlistmaker.search.data.network
 
 import android.content.Context
-import com.example.playlistmaker.data.dto.Response
-import com.example.playlistmaker.data.dto.TrackRequest
-import com.example.playlistmaker.util.NetworkUtils
+import com.example.playlistmaker.search.data.NetworkClient
+import com.example.playlistmaker.search.data.dto.Response
+import com.example.playlistmaker.search.data.dto.TrackRequest
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
