+++ app/src/main/java/com/example/playlistmaker/search/ui/TrackAdapter.kt
@@ -1,16 +1,15 @@
-package com.example.playlistmaker.presentation
+package com.example.playlistmaker.search.ui
 import android.view.LayoutInflater
+import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
-import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.example.playlistmaker.R
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.search.domain.models.Track
 import java.text.SimpleDateFormat
-import java.util.*
-import kotlin.collections.ArrayList
+import java.util.Locale
 
 class TrackAdapter(val clickListener: TrackClickListener): RecyclerView.Adapter<TrackAdapter.TrackHolder>() {
     var tracks = ArrayList<Track>()
