+++ app/src/main/res/layout/activity_audio_player.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.AudioPlayerActivity">
+    tools:context=".player.ui.AudioPlayerActivity">
 
         <ImageButton
             android:id="@+id/back"
