+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:theme="@style/Main"
-    tools:context=".presentation.MainActivity"
+    tools:context=".main.MainActivity"
     >
