+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".presentation.SearchActivity"
+    tools:context=".search.ui.SearchActivity"
     android:background="?attr/whiteBlack"
    >
